+++ src/haddock/modules/refinement/emref/__init__.py
@@ -4,8 +4,8 @@ from os import linesep
 from pathlib import Path
 from haddock.modules import BaseHaddockModule
 from haddock.libs.libsubprocess import CNSJob
-from haddock.cns.util import generate_default_header, load_ambig
-from haddock.cns.util import load_workflow_params, prepare_multiple_input
+from haddock.libs.libcns.util import generate_default_header, load_ambig
+from haddock.libs.libcns.util import load_workflow_params, prepare_multiple_input
 from haddock.libs.libparallel import Scheduler
 from haddock.libs.libontology import Format, ModuleIO, PDBFile
