+++ src/haddock/modules/sampling/rigidbody/__init__.py
@@ -5,8 +5,8 @@ from pathlib import Path
 from haddock.gear.haddockmodel import HaddockModel
 from haddock.modules import BaseHaddockModule
 from haddock.libs.libsubprocess import CNSJob
-from haddock.cns.util import generate_default_header, load_ambig
-from haddock.cns.util import load_workflow_params, prepare_multiple_input
+from haddock.libs.libcns.util import generate_default_header, load_ambig
+from haddock.libs.libcns.util import load_workflow_params, prepare_multiple_input
 from haddock.libs.libparallel import Scheduler
 from haddock.libs.libontology import Format, ModuleIO, PDBFile
